♻️ [유경] refactor : post페이지 스타일 컴포넌트 분리
